[kernel] Build everything with -Wconversion [part 33]
Note: improved sort_test.cpp to assess the changes in kernel/sort.c.

Changes to be committed:
	modified:   include/tilck/kernel/sort.h
	modified:   kernel/mm/general_kmalloc.c.h
	modified:   kernel/mm/kmalloc_heaps.c.h
	modified:   kernel/mm/system_mmap.c
	modified:   kernel/sort.c
	modified:   unittests/sort_test.cpp

+++ kernel/mm/general_kmalloc.c.h
@@ -45,7 +45,7 @@ general_kmalloc(size_t *size, u32 flags)
    }
 
    // Iterate in reverse-order because the first heaps are the biggest ones.
-   for (int i = used_heaps - 1; i >= 0; i--) {
+   for (int i = (int)used_heaps - 1; i >= 0; i--) {
 
       ASSERT(heaps[i] != NULL);
 
@@ -97,7 +97,7 @@ general_kfree(void *ptr, size_t *size, u32 flags)
 
    kmalloc_heap *h = NULL;
 
-   for (int i = used_heaps - 1; i >= 0; i--) {
+   for (int i = (int)used_heaps - 1; i >= 0; i--) {
 
       uptr hva = heaps[i]->vaddr;
